Time: 0 ms (100.00%) | Memory: 23.6 MB (23.09%) - LeetSync

+++ 155-min-stack/min-stack.cpp
@@ -33,10 +33,10 @@ public:
     
     int top() {
        if(st.top()>=mn) return (int)st.top();
-       else return (int)mn;                       
+       else return mn;                       
     }
     
     int getMin() {
-        return (int)mn;
+        return mn;
     }
 };
